chore: small refactor and improvements to packages (#52)
## Description

We had started using the `@elba-security` namespace for all internal packages, but didn't migrate the config packages. In this PR we are refactoring the `tsconfig` and `eslint-config-custom` packages to be part of the `@elba-security` namespace. Also it features some additional tiny improvements.

## Checklist:

Before you create this PR, confirm all the requirements listed below by checking the checkboxes like this (`[x]`).

- [X] I have performed a self-review of my own code.
- [ ] I have commented my code, particularly in hard-to-understand areas.
- [ ] I have added tests to cover the new feature or fixes.