chore(sdk): add util functions to use on cryptographic validation of the elba signature (#53)
## Description

Add util function to the SDK to help with the validation of the cryptographic signature sent to the integration's code on the request expected to come from elba. The request will be signed using a secret know by elba and the integration, and also using the request's payload. The integratot should be able to use the `isValidWebhookElbaSignature` function to verify the authencity of the request.

## Checklist:

Before you create this PR, confirm all the requirements listed below by checking the checkboxes like this (`[x]`).

- [X] I have performed a self-review of my own code.
- [X] I have commented my code, particularly in hard-to-understand areas.
- [X] I have added tests to cover the new feature or fixes.